+++ RangeHighlight/Integrations.cs
@@ -5,27 +5,36 @@ using StardewModdingAPI;
 
 namespace RangeHighlight {
     internal class Integrations {
-        private IModHelper helper;
-        private IRangeHighlightAPI highlightApi;
-        private ModConfig config;
-        private ModEntry.DefaultShapes defaultShapes;
-        public Integrations(IModHelper helper, IRangeHighlightAPI api, ModConfig config, ModEntry.DefaultShapes defaultShapes) {
-            this.helper = helper;
-            this.highlightApi = api;
-            this.config = config;
-            this.defaultShapes = defaultShapes;
+        private ModEntry theMod;
+        public Integrations(ModEntry theMod) {
+            this.theMod = theMod;
             IntegratePrismaticTools();
+            IntegrateBetterJunimos();
         }
 
         private void IntegratePrismaticTools() {
-            IPrismaticToolsAPI api = helper.ModRegistry.GetApi<IPrismaticToolsAPI>("stokastic.PrismaticTools");
+            IPrismaticToolsAPI api = theMod.helper.ModRegistry.GetApi<IPrismaticToolsAPI>("stokastic.PrismaticTools");
             if (api == null) return;
-            defaultShapes.prismaticSprinkler = highlightApi.GetSquareCircle((uint)api.SprinklerRange);
+            theMod.defaultShapes.prismaticSprinkler = theMod.api.GetSquareCircle((uint)api.SprinklerRange);
+        }
+        private void IntegrateBetterJunimos() {
+            IBetterJunimosAPI api = theMod.helper.ModRegistry.GetApi<IBetterJunimosAPI>("hawkfalcon.BetterJunimos");
+            if (api == null) return;
+            int r = api.GetJunimoHutMaxRadius();
+            if (r > 1) {
+                theMod.defaultShapes.SetJunimoRange((uint)r);
+            } else {
+                theMod.Monitor.Log($"ignoring nonsense value {r} from Better Junimos for Junimo Hut radius", LogLevel.Info);
+            }
         }
     }
 
     public interface IPrismaticToolsAPI {
         int SprinklerRange { get; }
         int SprinklerIndex { get; }
+    }
+
+    public interface IBetterJunimosAPI {
+        int GetJunimoHutMaxRadius();
     }
 }